+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ## [Unreleased] <!-- ReleaseDate -->
 
-- Fix unexpected behaviour of `keep_filter` option in MultiSelect prompts, where the resulting behaviour was the opposite of what was expected.
+- Fix unexpected behaviors of `keep_filter` option in MultiSelect prompts:
+  - Filter input is now correcly getting reset **only when** `keep_filter == false`.
+  - When the filter input is reset, the list of options is reset as well. Thanks @Swivelgames for reporting [#238](https://github.com/mikaelmello/inquire/issues/238).
 
 ## [0.7.3] - 2024-03-21